+++ experimental/CHANGELOG.md
@@ -6,12 +6,15 @@ All notable changes to experimental packages in this project will be documented 
 
 ### :boom: Breaking Change
 
+* refactor(metrics-sdk): rename InstrumentationLibrary -> InstrumentationScope #2959 @pichlermarc
+
 ### :rocket: (Enhancement)
 
 * feat(exporters): update proto version and use otlp-transformer #2929 @pichlermarc
 
 ### :bug: (Bug Fix)
 
+* fix(opentelemetry-instrumentation-http): use correct origin when port is `null` #2948 @danielgblanco
 * fix(otlp-exporter-base): include esm and esnext in package files #2952 @dyladan
 
 ### :books: (Refine Doc)